refactor(backend): avoid code repetition
create shared types and graphql schemas, and create a custom repository with the method 'findMany' to avoid code repetition
