+++ backend/src/shared/types/Filter.ts
@@ -1,0 +1,6 @@
+export default interface Filter {
+  [key: string]: {
+    operator: '=' | '>' | '>=' | '<' | '<=' | '!=';
+    value: string | number;
+  };
+}
