+++ backend/src/shared/types/Order.ts
@@ -1,0 +1,3 @@
+export default interface Order {
+  [key: string]: 'ASC' | 'DESC';
+}
