+++ backend/src/shared/types/Pagination.ts
@@ -1,0 +1,4 @@
+export default interface Pagination {
+  page: number;
+  take: number;
+}
