+++ backend/src/shared/types/index.ts
@@ -1,0 +1,3 @@
+export { default as Filter } from './Filter';
+export { default as Order } from './Order';
+export { default as Pagination } from './Pagination';
